# clientgin/routes/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/item": {
            "get": {
                "description": "do ping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item"
                ],
                "summary": "ping example",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "position",
                        "name": "pos",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "count",
                        "name": "count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user_item.TItem"
                        }
                    }
                }
            },
            "post": {
                "description": "do ping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item"
                ],
                "summary": "ping example",
                "parameters": [
                    {
                        "description": "Thông tin Item mới ",
                        "name": "NewItem",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.RequestItem"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user_item.TItem"
                        }
                    }
                }
            }
        },
        "/item/{id}": {
            "put": {
                "description": "do ping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item"
                ],
                "summary": "ping example",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID cần sửa",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Thông tin cần sửa",
                        "name": "ItemEdit",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user_item.TItem"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user_item.TItem"
                        }
                    }
                }
            },
            "delete": {
                "description": "do ping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item"
                ],
                "summary": "ping example",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id cần xóa",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user_item.TItem"
                        }
                    }
                }
            }
        },
        "/item/{uid}": {
            "get": {
                "description": "do ping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item"
                ],
                "summary": "ping example",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID của người dùng cần tìm kiếm ",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "position",
                        "name": "pos",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "count",
                        "name": "count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user_item.TItem"
                        }
                    }
                }
            }
        },
        "/user": {
            "post": {
                "description": "do ping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "ping example",
                "parameters": [
                    {
                        "description": "Thông tin user mới ",
                        "name": "NewUser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user_item.TUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user_item.TUser"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.RequestItem": {
            "type": "object",
            "properties": {
                "arrOwnerAddress": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "arrOwnerUid": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "categoryType": {
                    "type": "integer"
                },
                "chain": {
                    "type": "string"
                },
                "collectionId": {
                    "type": "integer"
                },
                "contractAddress": {
                    "type": "string"
                },
                "creatorAddress": {
                    "type": "string"
                },
                "creatorUid": {
                    "type": "integer"
                },
                "currencyAddress": {
                    "$ref": "#/definitions/user_item.TCurrencyAddress"
                },
                "edition": {
                    "type": "integer"
                },
                "endAt": {
                    "type": "integer"
                },
                "eventIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "extendData": {
                    "$ref": "#/definitions/user_item.TExtendData"
                },
                "externalLink": {
                    "type": "string"
                },
                "hidden": {
                    "type": "boolean"
                },
                "imageUrl": {
                    "type": "string"
                },
                "isLootbox": {
                    "type": "boolean"
                },
                "isSensitive": {
                    "type": "boolean"
                },
                "mediaType": {
                    "type": "string"
                },
                "metaData": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "ownerAddress": {
                    "type": "string"
                },
                "ownerUID": {
                    "type": "integer"
                },
                "price": {
                    "type": "string"
                },
                "productNo": {
                    "type": "string"
                },
                "startAt": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "tokenId": {
                    "type": "string"
                },
                "totalClaim": {
                    "type": "integer"
                },
                "totalEdition": {
                    "type": "integer"
                },
                "totalLike": {
                    "type": "integer"
                },
                "totalLimit": {
                    "type": "integer"
                },
                "totalView": {
                    "type": "integer"
                },
                "typeNft": {
                    "type": "integer"
                },
                "unlockableContent": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "user_item.TCurrencyAddress": {
            "type": "object",
            "properties": {
                "Address": {
                    "type": "string"
                },
                "Chain": {
                    "type": "string"
                },
                "Decimal": {
                    "type": "integer"
                },
                "Img": {
                    "description": "unused field # 6",
                    "type": "string"
                },
                "Name": {
                    "type": "string"
                },
                "Status": {
                    "type": "boolean"
                },
                "Symbol": {
                    "type": "string"
                }
            }
        },
        "user_item.TExtendData": {
            "type": "object",
            "properties": {
                "Category": {
                    "type": "string"
                },
                "Collection": {
                    "type": "string"
                },
                "Create": {
                    "type": "string"
                },
                "Like": {
                    "type": "string"
                },
                "Owner": {
                    "type": "string"
                },
                "TotalOffer": {
                    "type": "string"
                }
            }
        },
        "user_item.TItem": {
            "type": "object",
            "properties": {
                "ArrOwnerAddress": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "ArrOwnerUid": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "CategoryType": {
                    "type": "integer"
                },
                "Chain": {
                    "type": "string"
                },
                "CollectionId": {
                    "type": "integer"
                },
                "ContractAddress": {
                    "type": "string"
                },
                "CreateTime": {
                    "type": "integer"
                },
                "CreatorAddress": {
                    "type": "string"
                },
                "CreatorUid": {
                    "type": "integer"
                },
                "CurrencyAddress": {
                    "$ref": "#/definitions/user_item.TCurrencyAddress"
                },
                "Edition": {
                    "type": "integer"
                },
                "EndAt": {
                    "type": "integer"
                },
                "EventIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "ExtendData": {
                    "$ref": "#/definitions/user_item.TExtendData"
                },
                "ExternalLink": {
                    "type": "string"
                },
                "Hidden": {
                    "type": "boolean"
                },
                "Id": {
                    "type": "integer"
                },
                "ImageUrl": {
                    "type": "string"
                },
                "IsLootbox": {
                    "type": "boolean"
                },
                "IsSensitive": {
                    "type": "boolean"
                },
                "MediaType": {
                    "type": "string"
                },
                "MetaData": {
                    "type": "string"
                },
                "Name": {
                    "type": "string"
                },
                "OwnerAddress": {
                    "type": "string"
                },
                "OwnerUID": {
                    "type": "integer"
                },
                "Price": {
                    "type": "string"
                },
                "ProductNo": {
                    "type": "string"
                },
                "StartAt": {
                    "type": "integer"
                },
                "Status": {
                    "type": "integer"
                },
                "TokenId": {
                    "type": "string"
                },
                "TotalClaim": {
                    "type": "integer"
                },
                "TotalEdition": {
                    "type": "integer"
                },
                "TotalLike": {
                    "type": "integer"
                },
                "TotalLimit": {
                    "type": "integer"
                },
                "TotalView": {
                    "type": "integer"
                },
                "TypeNft": {
                    "type": "integer"
                },
                "UnlockableContent": {
                    "type": "string"
                },
                "UpdateTime": {
                    "type": "integer"
                },
                "Url": {
                    "type": "string"
                }
            }
        },
        "user_item.TUser": {
            "type": "object",
            "properties": {
                "App": {
                    "type": "string"
                },
                "Avatar": {
                    "type": "string"
                },
                "Background": {
                    "type": "string"
                },
                "Bio": {
                    "type": "string"
                },
                "Block": {
                    "type": "boolean"
                },
                "BlockchainAddress": {
                    "type": "string"
                },
                "CanMint": {
                    "type": "boolean"
                },
                "CountryCode": {
                    "type": "string"
                },
                "CreateTime": {
                    "type": "integer"
                },
                "Deleted": {
                    "type": "boolean"
                },
                "Email": {
                    "type": "string"
                },
                "ExtendData": {
                    "$ref": "#/definitions/user_item.TExtendData"
                },
                "Facebook": {
                    "type": "string"
                },
                "Instagram": {
                    "type": "string"
                },
                "IsPartner": {
                    "type": "boolean"
                },
                "Language": {
                    "type": "string"
                },
                "Name": {
                    "type": "string"
                },
                "Pubkey": {
                    "type": "string"
                },
                "TotalFollers": {
                    "type": "integer"
                },
                "TotalFollowing": {
                    "type": "integer"
                },
                "Twitter": {
                    "type": "string"
                },
                "TypeUser": {
                    "type": "integer"
                },
                "Uid": {
                    "type": "integer"
                },
                "UpdateTime": {
                    "type": "integer"
                },
                "Username": {
                    "type": "string"
                },
                "Verify": {
                    "type": "boolean"
                },
                "Vip": {
                    "type": "integer"
                },
                "Youtube": {
                    "type": "string"
                }
            }
        }
    }
}